Add a thinking animation for all skills while they are working (#5)
* Add a thinking animation for all skills while they are working

The 'thinking' animation now occurs on a Mark 1 automatically
whenever a skill handler is started, unless it reacts within a
quarter second.  This give feedback to the user that the
system heard them and is processing.

* initialize new settings variable

in older versions 0.9.15 and under, this skill will throw a key error every time you use the wake word. This is due to the skill settings erroring out due to the backend code changes that dynamically added a "name" for the label field.

* Add a thinking animation for all skills while they are working

The 'thinking' animation now occurs on a Mark 1 automatically
whenever a skill handler is started, unless it reacts within a
quarter second.  This give feedback to the user that the
system heard them and is processing.

* Addressing review concerns, and typo fix

* After a rebase, fixing mismatched square and round brackets
* Added and try/except for a rare case