Eye animation clean-up
Due to the current implementation of the Arduino code on a Mark 1, at
times the eye transition animations get interrupted leaving the eyes
in an odd state.  This change performs a cleanup of the whole eye
when the eyes go to full idle.  So the problem might still occur, but
it disappears soon afterwards when it does.

Other minor changes:
* schedule_repeating_event() no longer requires a start time
* Support the EnclosureAPI implementation of eyes_setpixel()